Change health endpoint from ping to health
