+++ management_api_app/api/routes/api.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter
 
-from api.routes import ping, health
+from api.routes import health, status
 
 
 router = APIRouter()
-router.include_router(ping.router, tags=["ping"])
 router.include_router(health.router, tags=["health"])
+router.include_router(status.router, tags=["status"])
