+++ management_api_app/api/routes/health.py
@@ -1,14 +1,12 @@
+from datetime import datetime
 from fastapi import APIRouter
-from models.schemas.health import HealthCheck, ServiceStatus
+from models.schemas.health import Pong
 from resources import strings
-from services.health_checker import create_state_store_status
 
 
 router = APIRouter()
 
 
-@router.get("/health", name="health:get-status-of-services")
-async def health_check() -> HealthCheck:
-    status, message = create_state_store_status()
-    services = [ServiceStatus(service=strings.COSMOS_DB, status=status, message=message)]
-    return HealthCheck(services=services)
+@router.get("/health", name="health:get-server-alive")
+async def ping_server() -> Pong:
+    return Pong(message=strings.PONG, time=datetime.now())
