+++ templates/core/terraform/appgateway/appgateway.tf
@@ -63,7 +63,7 @@ resource "azurerm_application_gateway" "agw" {
     pick_host_name_from_backend_http_settings = true
     interval                                  = 10
     protocol                                  = "Https"
-    path                                      = "/api/ping"
+    path                                      = "/api/health"
     timeout                                   = 10
     unhealthy_threshold                       = 2
     minimum_servers                           = 0
